+++ spec/requests/admin_timeout_spec.rb
@@ -40,7 +40,7 @@ RSpec.describe "Admin session timing out", type: :request do
       expect(session[:role_codes]).to eq([AdminSession::SERVICE_OPERATOR_DFE_SIGN_IN_ROLE_CODE])
 
       travel after_expiry do
-        get root_path
+        get new_claim_path
 
         expect(session[:user_id]).to be_nil
         expect(session[:organisation_id]).to be_nil